Cleaning + test build
