+++ src/core/drag/dragDropManager.ts
@@ -1,5 +1,6 @@
 import type { SvelteRow } from '../row';
-import { type Writable, get } from 'svelte/store';
+import { get } from 'svelte/store';
+import type { Writable } from 'svelte/store';
 import type { EntityStore } from '../store';
 
 export type DropHandler = (event: MouseEvent) => any;
